feat: add Gates for authorization

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -15,6 +15,11 @@ class AttendeeController extends Controller
 
     private array $relations = ['user'];
 
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -66,6 +71,7 @@ class AttendeeController extends Controller
      */
     public function destroy(Event $event, Attendee $attendee)
     {
+        $this->authorize('delete-attendee', [$event, $attendee]);
         $attendee->delete();
 
         return response(status: 204);
